+++ src/components/layout/table/Table.stories.tsx
@@ -1,14 +1,27 @@
 import React from "react";
 
-import { Table } from "../../../../index";
+import type { StoryObj, Meta } from "@storybook/react";
 
-import styles from "./TableDemo.module.scss";
+import { Table } from "./Table";
 
-interface Props {}
+const meta: Meta = {
+    title: "Components/Layout/Table/Table",
+    component: Table,
+    tags: ["autodocs", "layout"],
 
-const TableDemo: React.FC<Props> = (props) => {
-    return (
-        <Table variant={"striped"}>
+    argTypes: {},
+    args: {
+        striped: true,
+        wide: true,
+        centered: true,
+    },
+};
+
+type Story = StoryObj<typeof Table>;
+
+const Primary: Story = {
+    render: args => (
+        <Table {...args}>
             <thead>
                 <tr>
                     <th>Name</th>
@@ -20,36 +33,39 @@ const TableDemo: React.FC<Props> = (props) => {
                 <tr>
                     <td>Stefan</td>
                     <td>41</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
                 <tr>
                     <td>Juliette</td>
                     <td>23</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
                 <tr>
                     <td>Amy</td>
                     <td>55</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
                 <tr>
                     <td>Stefan</td>
                     <td>41</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
                 <tr>
                     <td>Juliette</td>
                     <td>23</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
                 <tr>
                     <td>Amy</td>
                     <td>55</td>
-                    <td className={styles.actions}>edit</td>
+                    <td>edit</td>
                 </tr>
             </tbody>
         </Table>
-    );
+    ),
 };
 
-export { TableDemo };
+export default meta;
+export {
+    Primary,
+};